// cinepass-frontend/angular-front/src/app/interfaces/updateShowDTO.ts
export interface UpdateShowDto {
  dateAndTime?: Date; // UTC Fecha y hora de la función
  movie?: number;  // ID de la película
  showType?: number; // ID del tipo de show
  selectedLanguage?: number; // ID del idioma
  room?: number; // ID de la sala
  subsidiary?: number; // ID de la subsidiaria
}